requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
nodriver
pydantic==2.9.2
aiofiles==24.1.0
python-multipart==0.0.12
aiohttp>=3.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
